+++ opentelemetry-user-events-metrics/examples/basic-metrics.rs
@@ -59,104 +59,92 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .f64_observable_gauge("observable_gauge_f64_test")
         .with_unit("test_unit")
         .with_description("test_description")
+        .with_callback(|observer| {
+            observer.observe(
+                1.0,
+                &[
+                    KeyValue::new("mykey1", "myvalue1"),
+                    KeyValue::new("mykey2", "myvalue2"),
+                ],
+            )
+        })
         .init();
 
     let gauge2 = meter
         .u64_observable_gauge("observable_gauge_u64_test")
         .with_unit("test_unit")
         .with_description("test_description")
+        .with_callback(|observer| {
+            observer.observe(
+                1,
+                &[
+                    KeyValue::new("mykey1", "myvalue1"),
+                    KeyValue::new("mykey2", "myvalue2"),
+                ],
+            )
+        })
         .init();
-
-    meter.register_callback(&[gauge.as_any()], move |observer| {
-        observer.observe_f64(
-            &gauge,
-            1.0,
-            &[
-                KeyValue::new("mykey1", "myvalue1"),
-                KeyValue::new("mykey2", "myvalue2"),
-            ],
-        )
-    })?;
-
-    meter.register_callback(&[gauge2.as_any()], move |observer| {
-        observer.observe_u64(
-            &gauge2,
-            1,
-            &[
-                KeyValue::new("mykey1", "myvalue1"),
-                KeyValue::new("mykey2", "myvalue2"),
-            ],
-        )
-    })?;
 
     // Create a ObservableCounter instrument and register a callback that reports the measurement.
     let observable_counter = meter
         .u64_observable_counter("observable_counter_u64_test")
         .with_description("test_description")
         .with_unit("test_unit")
+        .with_callback(|observer| {
+            observer.observe(
+                100,
+                &[
+                    KeyValue::new("mykey1", "myvalue1"),
+                    KeyValue::new("mykey2", "myvalue2"),
+                ],
+            )
+        })
         .init();
 
     let observable_counter2 = meter
         .f64_observable_counter("observable_counter_f64_test")
         .with_description("test_description")
         .with_unit("test_unit")
+        .with_callback(|observer| {
+            observer.observe(
+                100.0,
+                &[
+                    KeyValue::new("mykey1", "myvalue1"),
+                    KeyValue::new("mykey2", "myvalue2"),
+                ],
+            )
+        })
         .init();
-
-    meter.register_callback(&[observable_counter.as_any()], move |observer| {
-        observer.observe_u64(
-            &observable_counter,
-            100,
-            &[
-                KeyValue::new("mykey1", "myvalue1"),
-                KeyValue::new("mykey2", "myvalue2"),
-            ],
-        )
-    })?;
-
-    meter.register_callback(&[observable_counter2.as_any()], move |observer| {
-        observer.observe_f64(
-            &observable_counter2,
-            100.0,
-            &[
-                KeyValue::new("mykey1", "myvalue1"),
-                KeyValue::new("mykey2", "myvalue2"),
-            ],
-        )
-    })?;
 
     // Create a Observable UpDownCounter instrument and register a callback that reports the measurement.
     let observable_up_down_counter = meter
         .i64_observable_up_down_counter("observable_up_down_counter_i64_test")
         .with_description("test_description")
         .with_unit("test_unit")
+        .with_callback(|observer| {
+            observer.observe(
+                100,
+                &[
+                    KeyValue::new("mykey1", "myvalue1"),
+                    KeyValue::new("mykey2", "myvalue2"),
+                ],
+            )
+        })
         .init();
     let observable_up_down_counter2 = meter
         .f64_observable_up_down_counter("observable_up_down_counter_f64_test")
         .with_description("test_description")
         .with_unit("test_unit")
+        .with_callback(|observer| {
+            observer.observe(
+                100.0,
+                &[
+                    KeyValue::new("mykey1", "myvalue1"),
+                    KeyValue::new("mykey2", "myvalue2"),
+                ],
+            )
+        })
         .init();
-
-    meter.register_callback(&[observable_up_down_counter.as_any()], move |observer| {
-        observer.observe_i64(
-            &observable_up_down_counter,
-            100,
-            &[
-                KeyValue::new("mykey1", "myvalue1"),
-                KeyValue::new("mykey2", "myvalue2"),
-            ],
-        )
-    })?;
-
-    meter.register_callback(&[observable_up_down_counter2.as_any()], move |observer| {
-        observer.observe_f64(
-            &observable_up_down_counter2,
-            100.0,
-            &[
-                KeyValue::new("mykey1", "myvalue1"),
-                KeyValue::new("mykey2", "myvalue2"),
-            ],
-        )
-    })?;
 
     loop {
         // Record measurements using the Counter instrument.