+++ opentelemetry-datadog/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## vNext
 
+## v0.13.0
+
+- Bump opentelemetry and opentelemetry_sdk version to 0.24
+
 ## v0.12.0
 
 ### Changed